NMS-13357: Lever regex for indexing of resource paths (resorceId): remove index tags
